Fix pyTorch dtype

+++ src/_torch_cpu.py
@@ -1,4 +1,4 @@
 import torch
 def rootloop(N):
-	# Coerce to float as pyTorch has not defined square root for ints
-	return torch.sum( torch.sqrt( torch.arange(0.0, N, 2, device=torch.device('cpu')) ) )
+	# Coerce to double as pyTorch defaults to float32 otherwise
+	return torch.sum( torch.sqrt( torch.arange(0, N, 2, device=torch.device('cpu'), dtype=torch.double) ) )
